opt: Move join-related custom funcs to new file
Split join-related custom funcs to new file, as the amount of join-
related code is steadily increasing.

Also update join test case to make an additional column not-null for
testing purposes.

Release note: None
